requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.29.0
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ecdsa==0.19.0
fastapi==0.112.2
greenlet==3.0.3
h11==0.14.0
idna==3.8
iniconfig==2.0.0
kombu==5.4.0
minio==7.2.8
packaging==24.1
pillow==10.4.0
pip==23.2.1
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.9
redis==5.0.8
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
tzdata==2024.1
unicorn==2.0.1.post1
urllib3==2.2.2
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
alembic==1.13.2